Created Java/relationChecker for !1 ontology
Added Java snippet (relationChecker) using OWLAPI and a number of
things. At this moment it checks the relations for concepts expressed in
the Manchester syntax, which I consider to be a fairly suitable syntax
for the intended manual use. Next step consists on considering a set of
ontologies.

Special thanks to Luis Manuel Vilches for the ontology for the tests.