relationChecker finished
Not many lines but quite a few features (OWLAPI is that powerful). This
snippet parses properly Manchester syntax names and when this fails it
parses full URIs, several ontologies can be added and removed, even with
conflicting short names. If there are conflicting short names the full
URI should be used to avoid ambiguity. This should never happen, but
it's covered just in case.

Minor corrections may happen in the future but the features I wanted to
test are already tested, so I expect not to work further on this
snippet.