Hanoi towers
Added a some code for hanoi towers (3 pegs) for a "kata", lazy version
and "normal" version. It's interesting because it's not possible to
reuse the same generator in the lazy version, I may try this same kata
with different languages to test their lazy and functional capabilities.

Also notice that I'm not passing the parameters down, but _up_, which
means that **partial results can be reused**. I'm using templates to do
some kind of "variable unification", suggestions about this are welcome,
it _feels_ like it could be much better but I cannot find the way at
this moment (may try next weekend again).