Considering mereologies in relationChecker
The support for mereologies is partial (very partial), but that's all
what makes sense in the current context. In this case relations like X
subclassOf partOf some Y do not make sense, since anything that is
necessarily part of something bigger would not be a first class citizen
and would not be considered for relations, thus, we only need to get the
contents of some aggregate considering those contents may (also) exist
without a container.