use REST from js, small example
If you want to test a REST service using PUT, GET, POST and DELETE
methods all you need to do is set the URL in line 28 and then use the
examples in lines 32-35. Please note that in the case of POST you may
need to introduce some data (parameter data), you can combine all these
calls in any preferred way. This is a minimum wrapper, better things
could be done, but that would take more time.