Updated testers for the REST service
In Python and Javascript, now both working with the new API.

The API should not have changed IMHO. Be aware that the Javascript test is much smaller than the Python test, I simply took one call from the Python test and put it in the Javascript test. In Python random values are generated.